Revert "Separate viewmodel for FindPasswordFragment.kt"
